refactor: Layers, Components 디렉토리 구조화 [#98]
- RoadMap 을 Layers 폴더로 분리
