+++ client/src/pages/WorkPage.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 
 import SideBar from '@/components/SideBar';
 import SideBarEntry from '@/components/SideBarEntry';
-import Roadmap from '@/components/Roadmap';
+import Roadmap from '@/layers/Roadmap';
 import Kanban from '@/layers/Kanban';
 import Backlog from '@/components/Backlog';
 import useTabs from '@/lib/hooks/useTabs';
